// src/app/estados.enums.ts
export enum estadosA1 {
  q0 = 0,
  q1 = 1,
  q2 = 2,
  q3 = 3
}

export enum estadosA2 {
  q0 = 0,
  q1 = 1,
  q2 = 2,
  q3 = 3
}

export enum estadosA3 {
  q0 = 0,
  q1 = 1,
  q2 = 2,
  q3 = 3,
  q4 = 4,
  q5 = 5,
  q6 = 6,
  q7 = 7,
  q8 = 8,
  q9 = 9
}

export enum estadosA4 {
  q0 = 0,
  q1 = 1,
  q2 = 2,
  q3 = 3,
  q4 = 4,
}


export enum estadosA5 {
  q0 = 0,
  q1 = 1,
  q2 = 2,
  q3 = 3
}